Moves file_channel out to its own files
